refactor: split the yt-dlp mappers behind an unchanged barrel
